dox:adminAPI 통합
