+++ src/api/adminAPI/adminAPI.types.ts
