--- src/api/adminReviewAPI/adminReviewAPI.ts
@@ -1,7 +1,0 @@
-import apiFetch from '@/api/common';
-
-const getAllReviews = async () => {
-  return await apiFetch.get(`/admin/reviews`);
-};
-
-export { getAllReviews };
